Reorganize types for met norway
